refactor(input): split stationInput god file into focused modules
stationInput.ts (1219 LOC) becomes a ~300-LOC composition root. Extracts:
- sequenceNormalize.ts: byte normalization + cursor-key/soft-newline helpers
- stationRows.ts: snapshot row-query helpers (store passed explicitly, DRY'd onto findWorktreeRowById)
- executeOutcome.ts: the RouteOutcome dispatcher + context-menu action resolution
- managedLaunch.ts: createManagedLaunch sub-factory (observer launch/attach/focus/new-session lifecycle); folds the 4x landInPane->closeOverlay->ack tail into landOnPane
- paneEffects.ts: createPaneEffects sub-factory (terminal IO, pane lifecycle, automations, sendWhenReady); owns splitSeq + nextSplitSeqFromPanes; adds a depth guard to worktreeIdForPane

The commands<->effects closure cycle is dissolved: sub-factories depend on options/stores, not the effects object. normalizeSequence/nextSplitSeqFromPanes/executeOutcome re-exported so app + tests are untouched. Unit describe blocks colocated into sequenceNormalize.test.ts + paneEffects.test.ts. Typecheck clean; input dir 179 pass / 0 fail.
